bulk-migrate-lulu/requirements/base.txt
# SHA1:c487ee8d0124e1e93dedea0f66536e670f521206
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
cachetools==5.2.1
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
google-api-core[grpc]==2.11.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.16.0
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-bigquery==3.4.1
    # via -r requirements/base.in
google-cloud-core==2.3.2
    # via google-cloud-bigquery
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.0
    # via google-cloud-bigquery
googleapis-common-protos==1.58.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.51.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.51.1
    # via google-api-core
idna==3.4
    # via requests
packaging==21.3
    # via google-cloud-bigquery
proto-plus==1.22.2
    # via google-cloud-bigquery
protobuf==4.21.12
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.10.4
    # via -r requirements/base.in
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via google-cloud-bigquery
python-dotenv==0.21.0
    # via -r requirements/base.in
python-gitlab==3.12.0
    # via -r requirements/base.in
requests==2.28.2
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   python-gitlab
    #   requests-toolbelt
requests-toolbelt==0.10.1
    # via python-gitlab
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.14
    # via requests
